+++ blade-service/fee/src/main/java/org/springblade/fee/service/impl/AlipayServiceImpl.java
@@ -10,6 +10,7 @@ import org.springblade.common.utils.AlipayConfig;
 import org.springblade.fee.service.AlipayService;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 
 
 @Service
@@ -26,7 +27,7 @@ public class AlipayServiceImpl  implements AlipayService {
 
 
 	@Override
-	public String webPagePay(String outTradeNo, Integer totalAmount, String subject) throws Exception {
+	public String webPagePay(String outTradeNo, BigDecimal totalAmount, String subject) throws Exception {
 
 		AlipayTradePagePayRequest alipayRequest = new AlipayTradePagePayRequest();
 		/** 同步通知，支付完成后，支付成功页面*/
